+++ include/boost/beast/http/impl/write.hpp
@@ -267,8 +267,8 @@ struct run_write_some_op
     void
     operator()(
         WriteHandler&& h,
-        Stream& s,
-        serializer<isRequest, Body, Fields>& sr)
+        Stream* s,
+        serializer<isRequest, Body, Fields>* sr)
     {
         // If you get an error on the following line it means
         // that your handler does not meet the documented type
@@ -283,9 +283,7 @@ struct run_write_some_op
             typename std::decay<WriteHandler>::type,
             Stream,
             isRequest, Body, Fields>(
-                std::forward<WriteHandler>(h),
-                s,
-                sr);
+                std::forward<WriteHandler>(h), *s, *sr);
     }
 };
 
@@ -299,9 +297,9 @@ struct run_write_op
     void
     operator()(
         WriteHandler&& h,
-        Stream& s,
+        Stream* s,
         Predicate const&,
-        serializer<isRequest, Body, Fields>& sr)
+        serializer<isRequest, Body, Fields>* sr)
     {
         // If you get an error on the following line it means
         // that your handler does not meet the documented type
@@ -317,9 +315,7 @@ struct run_write_op
             Stream,
             Predicate,
             isRequest, Body, Fields>(
-                std::forward<WriteHandler>(h),
-                s,
-                sr);
+                std::forward<WriteHandler>(h), *s, *sr);
     }
 };
 
@@ -333,8 +329,8 @@ struct run_write_msg_op
     void
     operator()(
         WriteHandler&& h,
-        Stream& s,
-        message<isRequest, Body, Fields>& m,
+        Stream* s,
+        message<isRequest, Body, Fields>* m,
         std::false_type,
         Args&&... args)
     {
@@ -351,9 +347,7 @@ struct run_write_msg_op
             typename std::decay<WriteHandler>::type,
             Stream,
             isRequest, Body, Fields>(
-                std::forward<WriteHandler>(h),
-                s,
-                m,
+                std::forward<WriteHandler>(h), *s, *m,
                 std::forward<Args>(args)...);
     }
 
@@ -365,8 +359,8 @@ struct run_write_msg_op
     void
     operator()(
         WriteHandler&& h,
-        Stream& s,
-        message<isRequest, Body, Fields> const& m,
+        Stream* s,
+        message<isRequest, Body, Fields> const* m,
         std::true_type,
         Args&&... args)
     {
@@ -383,9 +377,7 @@ struct run_write_msg_op
             typename std::decay<WriteHandler>::type,
             Stream,
             isRequest, Body, Fields>(
-                std::forward<WriteHandler>(h),
-                s,
-                m,
+                std::forward<WriteHandler>(h), *s, *m,
                 std::forward<Args>(args)...);
     }
 };
@@ -483,8 +475,8 @@ async_write_some_impl(
         void(error_code, std::size_t)>(
             run_write_some_op{},
             handler,
-            stream,
-            sr);
+            &stream,
+            &sr);
 }
 
 } // detail
@@ -638,9 +630,9 @@ async_write_header(
         void(error_code, std::size_t)>(
             detail::run_write_op{},
             handler,
-            stream,
+            &stream,
             detail::serializer_is_header_done{},
-            sr);
+            &sr);
 }
 
 //------------------------------------------------------------------------------
@@ -712,9 +704,9 @@ async_write(
         void(error_code, std::size_t)>(
             detail::run_write_op{},
             handler,
-            stream,
+            &stream,
             detail::serializer_is_done{},
-            sr);
+            &sr);
 }
 
 //------------------------------------------------------------------------------
@@ -834,8 +826,8 @@ async_write(
         void(error_code, std::size_t)>(
             detail::run_write_msg_op{},
             handler,
-            stream,
-            msg,
+            &stream,
+            &msg,
             std::false_type{});
 }
 
@@ -864,8 +856,8 @@ async_write(
         void(error_code, std::size_t)>(
             detail::run_write_msg_op{},
             handler,
-            stream,
-            msg,
+            &stream,
+            &msg,
             std::true_type{});
 }